msr_document_parser/requirements.txt
click==7.1.2
python-poppler==0.2.2
